requirements.txt
streamlit
pinecone-client
pinecone-text
langchain-community
langchain-huggingface
python-dotenv
